backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.36
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0
python-multipart==0.0.6
numpy==2.1.3
pandas==2.2.3
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
reportlab==4.0.7
openpyxl==3.1.2
<<<<<<< Updated upstream
psycopg[binary]==3.2.3
=======
psycopg[binary]==3.2.3
supabase==2.0.3
httpx>=0.24,<0.26
websockets>=11.0,<12.0
>>>>>>> Stashed changes
